+++ plugins/allure-wakamiti-plugin/src/main/java/es/iti/wakamiti/report/allure/internal/TestResult.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2022-2026 Instituto Tecnológico de Informática (ITI)
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+package es.iti.wakamiti.report.allure.internal;
+
+
+public class TestResult extends io.qameta.allure.model.TestResult implements WithUuid {
+
+}